Add LinkContainer.js
